fix DISPLAY_CFG0 for PewPew M4
This board was originally prototyped with a different display than what ended
up being sold. Both values are correct, depending on which display is used -
the difference is in the horizontal offset. The value 0x20180 is correct for
the board as it was sold, so that is likely what users want.

The value can be changed with the uf2patcher tool if needed.